refactor handling sarif, describe better
